Start reorganizing to separate rotation engine from API
