+++ py-packages/engine/corelle/engine/query/active-plates-at-time.sql
