+++ py-packages/engine/corelle/engine/query/plate-for-point.sql
