+++ py-packages/engine/corelle/engine/query/plates-for-model.sql
