+++ py-packages/engine/corelle/engine/query/rotation-pairs-for-model.sql
