+++ py-packages/engine/corelle/engine/query/rotation-pairs-for-time.sql
