+++ py-packages/engine/corelle/engine/query/rotation-pairs.sql
