+++ py-packages/engine/pyproject.toml
@@ -2,7 +2,7 @@
 authors = ["Daven Quinn <dev@davenquinn.com>"]
 description = "An API server application for the Corelle plate-rotation system"
 license = "MIT"
-name = "corelle.server"
+name = "corelle.rotate"
 version = "1.0.0"
 
 packages = [
@@ -11,24 +11,24 @@ packages = [
 
 [tool.poetry.dependencies]
 
-Flask = "^1"
-Flask-RESTful = "^0.3.9"
-MarkupSafe = "2.0.1"
+Fiona = "^1.8.22"
+IPython = "^8"
+PyYAML = "^6.0"
 "corelle.client" = {path = "../client", develop = true}
-"corelle.engine" = {path = "../engine", develop = true}
-itsdangerous = "<2.1.0"
-jinja2 = "<3.1.0"
 "macrostrat.database" = "2.0.0.dev3"
 "macrostrat.utils" = "^1.0.1"
+numba = "^0.56.3"
 numpy = "^1.23.4"
+numpy-quaternion = "^2022.4.2"
+psycopg2-binary = "^2.9.5"
 python = "^3.8"
-simplejson = "^3.17.6"
-"werkzeug" = "2.0.3"
+rich = "^12.6.0"
+typer = "^0.6.1"
 
 [tool.poetry.dev-dependencies]
 
 [tool.poetry.scripts]
-corelle = 'corelle.server.cli:cli'
+corelle = 'corelle.rotate.cli:cli'
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
